finished all users business logic with basic authentication(guard) for super admin and exception filter for validation

+++ src/features/users/users.service.ts
@@ -1,16 +1,9 @@
-import {
-  BadRequestException,
-  Injectable,
-  NotFoundException,
-} from "@nestjs/common";
-import { InjectRepository } from "@nestjs/typeorm";
-import { Repository } from "typeorm";
+import { Injectable } from "@nestjs/common";
 import { CreateUserDto } from "./dto/create-user.dto";
 import { User } from "./user.entity";
 import { PasswordService } from "../../applications/password.service";
 import { UsersCommandRepo } from "./repositories/users.commandRepo";
 import { UsersQueryRepo } from "./repositories/users.queryRepo";
-import { QueryInputUserDto } from "./dto/query-input-blog.dto";
 
 @Injectable()
 export class UsersService {
@@ -29,14 +22,14 @@ export class UsersService {
     user.createdAt = new Date();
     user.isConfirmed = false;
 
-    return this.usersCommandRepo.create(user);
-  }
-
-  async findAll(sortData: QueryInputUserDto) {
-    return await this.usersQueryRepo.find(sortData);
+    return await this.usersCommandRepo.create(user);
   }
 
   async remove(id: number) {
+    const findUser = await this.usersQueryRepo.findOne(id);
+    if (!findUser) {
+      return;
+    }
     return await this.usersCommandRepo.delete(id);
   }
 }
